管理后台11-07

+++ src/components/Channel.vue
@@ -2,56 +2,76 @@
   <div id="Channel">
     <el-table :data="tableData" border style="width:100%" @cell-click="cellActive">
       <el-table-column prop="no" label="通道编号" width="150">
-        <template scope="scope">
-          <input type="text" @change="inputChange(scope.$index)" v-model="scope.row.no">
+        <template slot-scope="scope">
+          <input type="text" @focus="inputChange(scope.$index)" @change="inputChange(scope.$index)"
+                 v-model="scope.row.no">
         </template>
       </el-table-column>
       <el-table-column prop="name" label="通道名称">
-        <template scope="scope">
-          <input type="text" @change="inputChange(scope.$index)" v-model="scope.row.name">
+        <template slot-scope="scope">
+          <input type="text" @focus="inputChange(scope.$index)" @change="inputChange(scope.$index)"
+                 v-model="scope.row.name">
         </template>
       </el-table-column>
       <el-table-column prop="callerName" label="调用者名称">
-        <template scope="scope">
-          <input type="text" @change="inputChange(scope.$index)" v-model="scope.row.callerName">
+        <template slot-scope="scope">
+          <input type="text" @focus="inputChange(scope.$index)" @change="inputChange(scope.$index)"
+                 v-model="scope.row.callerName">
         </template>
       </el-table-column>
-      <el-table-column prop="callerUrl" label="调用者回调地址">
-        <template scope="scope">
-          <input type="text" @change="inputChange(scope.$index)" v-model="scope.row.callerUrl">
+      <el-table-column prop="callerUrl" label="调用者回调地址" width="150">
+        <template slot-scope="scope">
+          <input type="text" @focus="inputChange(scope.$index)" @change="inputChange(scope.$index)"
+                 v-model="scope.row.callerUrl">
         </template>
       </el-table-column>
       <el-table-column prop="companyName" label="合作对象">
-        <template scope="scope">
-          <input type="text" @change="inputChange(scope.$index)" v-model="scope.row.companyName">
+        <template slot-scope="scope">
+          <!--<input type="text" :remote="!btnArr[scope.$index]"  remote  @focus="inputChange(scope.$index)" @change="inputChange(scope.$index)" v-model="scope.row.companyName">-->
+          <el-select filterable :remote="!btnArr[scope.$index]" reserve-keyword :remote-method="getcompanyName"
+                     @focus="inputChange(scope.$index)" @change="inputChange(scope.$index)"
+                     v-model="scope.row.companyName">
+            <el-option v-for="item in collaborate" :key="item.value" :label="item.label"
+                       :value="item.label"></el-option>
+          </el-select>
         </template>
       </el-table-column>
       <el-table-column prop="summary" label="通道用途">
-        <template scope="scope">
-          <input type="text" @change="inputChange(scope.$index)" v-model="scope.row.summary">
+        <template slot-scope="scope">
+          <input type="text" @focus="inputChange(scope.$index)" @change="inputChange(scope.$index)"
+                 v-model="scope.row.summary">
         </template>
       </el-table-column>
       <el-table-column prop="rate" label="通道费率">
-        <template scope="scope">
-          <input type="text" @change="inputChange(scope.$index)" v-model="scope.row.rate">
+        <template slot-scope="scope">
+          <input type="text" @focus="inputChange(scope.$index)" @change="inputChange(scope.$index)"
+                 v-model="scope.row.rate">
         </template>
       </el-table-column>
       <el-table-column prop="enableFlag" label="通道状态">
         <!--有效标志 0为无效 1为有效-->
-        <template scope="scope">
-          <input type="text" @change="inputChange(scope.$index)" v-model="scope.row.enableFlag">
+        <template slot-scope="scope">
+          <!--<input type="text" @focus="inputChange(scope.$index)" @change="inputChange(scope.$index)" v-model="scope.row.enableFlag">-->
+          <el-select filterable v-model="scope.row.enableFlag" @change="selectChange(scope.$index)">
+            <el-option v-for="item in enableFlag" :key="item.value" :label="item.label" :value="item.value"></el-option>
+          </el-select>
         </template>
       </el-table-column>
       <el-table-column prop="businessType" label="对应订单类型">
-        <template scope="scope">
-          <input type="text" @change="inputChange(scope.$index)" v-model="scope.row.businessType">
+        <template slot-scope="scope">
+          <!--<input type="text" @focus="inputChange(scope.$index)" @change="inputChange(scope.$index)" v-model="scope.row.businessType">-->
+          <el-select filterable v-model="scope.row.businessType" @change="selectChange(scope.$index)">
+            <el-option v-for="item in businessType" :key="item.value" :label="item.label"
+                       :value="item.value"></el-option>
+          </el-select>
         </template>
       </el-table-column>
-      <el-table-column label="操作">
-        <template scope="scope">
+      <el-table-column label="操作" width="180">
+        <template slot-scope="scope">
           <el-button
             size="small"
-            @click="handleEdit(scope.$index, scope.row , scope)"><span ref="btn">{{btnArr[scope.$index]}}</span>
+            @click="handleEdit(scope.$index, scope.row , scope)" :disabled="btnArr[scope.$index]"><span
+            ref="btn">保存</span>
           </el-button>
           <el-button
             size="small"
@@ -72,20 +92,76 @@
         tableData: [],
         inputValue: [],
         btnArr: [],
-        tdData: null
+        tdData: null,
+        collaborate: [],
+        enableFlag: [{
+          value: 0,
+          label: '不可用'
+        }, {
+          value: 1,
+          label: '可用'
+        }
+        ],
+        businessType: [{
+          value: 1,
+          label: '缴费'
+        }, {
+          value: 2,
+          label: '挂号'
+        }, {
+          value: 3,
+          label: '门诊'
+        }]
       }
     },
     methods: {
       addNewChannel(){
         this.tableData.push({channelId: (this.tableData.length + 1), newData: 1});
-        this.btnArr.push('编辑')
+        this.btnArr.push(true)
+      },
+      getCollaborate(name){
+        var vm = this;
+        vm.collaborate = [];
+        var collaborateGet = new RemoteCall();
+        collaborateGet.init({
+          router: '/company/get',
+          session: vm.session,
+          data: {
+            pageInfo: {
+              pageSize: 100,
+              pageNum: 1
+            },
+            name: name
+          },
+          callback: function (data) {
+            console.log(data)
+            if (data.rows.length == 0) {
+              vm.collaborate = [];
+            }
+            for (var i = 0; i < data.rows.length; i++) {
+              var options = {};
+              options.value = data.rows[i].id;
+              options.label = data.rows[i].name;
+              vm.$set(vm.collaborate, i, options);
+            }
+          }
+        })
       },
       inputChange(index){
 //          this.tableData[index].btn='保存';
-        this.$set(this.btnArr, index, '保存')
+        this.$set(this.btnArr, index, false)
         this.$nextTick(function () {
-          console.log(this.tableData);
+//          console.log(this.tableData);
         })
+      },
+      selectChange(index){
+        this.$set(this.btnArr, index, false);
+//         if(this.tableData[index].companyName=""){
+//           this.tableData[index].companyName=null;
+//         }
+//         if(this.tableData[index].companyName){
+//           this.getCollaborate(this.tableData[index].companyName)
+//         }
       },
       cellActive(row){
         if (row.channelId > 20) { //20代表请求回来的条数
@@ -102,8 +178,21 @@
           session: this.session,
           data: str,
           callback: function (data) {
-            console.log(data);
-            vm.checkChannel();
+            if (data.ret.errorCode === 0) {
+              vm.$alert('保存成功', '提示', {
+                confirmButtonText: '确定',
+                callback: function () {
+                  vm.checkChannel()
+                }
+              });
+            } else {
+              vm.$alert(data.ret.errorMessage, '提示', {
+                confirmButtonText: '确定',
+                callback: function () {
+                  vm.checkChannel()
+                }
+              });
+            }
           }
         })
       },
@@ -118,25 +207,35 @@
           session: this.session,
           data: str,
           callback: function (data) {
-            console.log(data);
+            vm.$alert('保存成功', '提示', {
+              confirmButtonText: '确定'
+            });
             vm.checkChannel();
+            vm.getCollaborate(null);
           }
         })
       },
-      deleteChannel(str){
+      deleteChannel(str){ //删除
         var vm = this;
-        var addChannel = new RemoteCall();
-        addChannel.init({
-          router: '/base/RouteApi/delete',
-          session: this.session,
-          data: {
-            id: str
-          },
-          callback: function (data) {
-            console.log(data);
-            vm.checkChannel();
-          }
-        })
+        if (str) {
+          var addChannel = new RemoteCall();
+          addChannel.init({
+            router: '/base/RouteApi/delete',
+            session: this.session,
+            data: {
+              id: str
+            },
+            callback: function (data) {
+              vm.$alert('删除成功', '提示', {
+                confirmButtonText: '确定'
+              });
+              vm.checkChannel();
+            }
+          })
+        } else {
+          vm.checkChannel();
+        }
+
       },
       checkChannel(){//查询接口内容
         var vm = this;
@@ -154,25 +253,14 @@
             if (data.rows) {
               vm.tableData = data.rows;
               for (var i = 0; i < vm.tableData.length; i++) {
-                vm.btnArr[i] = '编辑'
-                if (vm.tableData[i].enableFlag == 1) {
-                  vm.tableData[i].enableFlag = '可用'
-                } else {
-                  vm.tableData[i].enableFlag = '不可用'
-                }
-                switch (vm.tableData[i].businessType) {
-                  case 1:
-                    vm.tableData[i].businessType = '缴费';
-                    break;
-                  case 2:
-                    vm.tableData[i].businessType = '挂号';
-                    break;
-                  case 3:
-                    vm.tableData[i].businessType = '门诊';
-                    break;
-                }
+                vm.btnArr[i] = true
                 vm.tableData[i].rate = vm.tableData[i].rate * 100 + '%';
               }
+            }
+          },
+          errorCallback: function (data) {
+            if (data) {
+              vm.$router.push('/login')
             }
           }
         })
@@ -197,31 +285,37 @@
       handleEdit(index, row, data){
         row.companyId = this.companyId(row.companyName);
         if (row.rate) {
-          row.rate = parseInt(row.rate) / 100;
-        }
-        if (row.enableFlag == '可用') {
-          row.enableFlag = 1;
-        } else if (row.enableFlag == '不可用') {
-          row.enableFlag = 0;
-        } else {
-          alert('通道状态只能填写“可用”或“不可用”')
-        }
+          if (isNaN(parseInt(row.rate))) {
+            this.$alert('请输入对应的比例比如：10%', '提示', {
+              confirmButtonText: '确定'
+            });
+            row.rate = null;
+            return;
+          } else {
+            row.rate = parseInt(row.rate) / 100;
+          }
 
-        switch (row.businessType) {
-          case '缴费':
-            row.businessType = 1;
-            break;
-          case '挂号':
-            row.businessType = 2;
-            break;
-          case '门诊':
-            row.businessType = 3;
-            break;
-          default :
-            alert('对应订单类型只能填写 “缴费” ，“挂号” ，“门诊”')
         }
         if (row.companyId == null) {
-          alert('合作公司不存在，请添加');
+          row.rate = row.rate * 100 + '%';
+          if (row.enableFlag == 1) {
+            row.enableFlag = '可用';
+          } else if (row.enableFlag === 0) {
+            row.enableFlag = '不可用';
+          }
+          ;
+          if (row.businessType == 1) {
+            row.businessType = '缴费';
+          }
+          else if (row.businessType == 2) {
+            row.businessType = '挂号';
+          }
+          else if (row.businessType == 3) {
+            row.businessType = '门诊';
+          }
+          this.$alert('合作公司不存在，请添加', '提示', {
+            confirmButtonText: '确定',
+          });
           return
         }
         if (row.newData == 1) {//如果数据有newData标志调用添加接口
@@ -234,11 +328,32 @@
       },
       handleDelete(index, row) {
 //        console.log(row.id);
-        this.deleteChannel(row.id);
+        this.$confirm('此操作将永久删除该文件, 是否继续?', '提示', {
+          confirmButtonText: '确定',
+          cancelButtonText: '取消',
+          type: 'warning'
+        }).then(() => {
+
+          this.deleteChannel(row.id);
+
+        }).catch(() => {
+
+
+        });
+
+      },
+      getcompanyName(data){
+//        this.collaborate = [];
+        if (data != "") {
+          this.getCollaborate(data);
+        } else {
+          this.getCollaborate(null);
+        }
       }
     },
     mounted: function () {
-      this.checkChannel()
+      this.checkChannel();
+      this.getCollaborate(null)
     }
   }
 </script>
